Referential LWR (length weight relationship) table

Conversion of the samples length measurements in weight by length weight relationship (LWR).
Formula associated are the following one:
\(RWT=a \times LF^b\) where:

RWT

is the round weight (kg)

LF

is the curved fork length (cm)

a, b

parameters from the reference table, dependent of of the species and potentially of the area (ocean or others) and the season.

data(length_weight_relationship, package="t3")

Format

length_weight_relationship: data frame with 40 rows and 7 columns:

ocean_code

Ocean code, type integer. For example ocean_code=1 for the Atlantic Ocean and ocean_code=2 the Indian Ocean

ocean_label

Ocean label ("Atlantic" or "Indian"), type character.

species_code

Species code, type integer.

species_fao_code

Species FAO code, type character.

length_weight_formula

LWR formula form, type character.

lwr_a

Parameter a of LWR formula, type double.

lwr_b

Parameter b of LWR formula, type double.

Details

More detail information could be find on the regional fisheries management organisations (RFMOs) like ICCAT or IOTC.