Reference table defining the logbook program for Observe database to be considered in T3 process.

Reference table defining the logbook program for Observe database to be considered in T3 process.

data(observe_logbook_program, package="t3")

Format

observe_logbook_program: data frame with 29 rows and 3 columns:

topiaid: logbook program topiaid, type character.

code: program code, type character.

label1: program label in English, type numeric.

Examples

data(observe_logbook_program, package="t3")
 observe_logbook_program
#>                                                                    topiaid
#> 1    fr.ird.referential.ps.common.Program#1686418523723#0.6779107415087587
#> 2  fr.ird.referential.ps.common.Program#1712910028020#0.042305269738915796
#> 3   fr.ird.referential.ps.common.Program#1712909866755#0.14694416632125606
#> 4      fr.ird.referential.ps.common.Program#1687453427699#0.45600561446958
#> 5    fr.ird.referential.ps.common.Program#1708965104274#0.9360663761762511
#> 6  fr.ird.referential.ps.common.Program#1686418198061#0.028969129326742893
#> 7     fr.ird.referential.ps.common.Program#1708965054280#0.628800018051336
#> 8    fr.ird.referential.ps.common.Program#1686418563724#0.3458091645701752
#> 9    fr.ird.referential.ps.common.Program#1726732909719#0.5539768483017606
#> 10  fr.ird.referential.ps.common.Program#1726492082043#0.07996395751781671
#> 11   fr.ird.referential.ps.common.Program#1712909995215#0.9034246862669405
#> 12   fr.ird.referential.ps.common.Program#1686418633374#0.6091104168659148
#> 13   fr.ird.referential.ps.common.Program#1726043223649#0.7009857103663292
#> 14   fr.ird.referential.ps.common.Program#1726492368048#0.5164681507213593
#> 15   fr.ird.referential.ps.common.Program#1731686160256#0.2218628119719901
#> 16   fr.ird.referential.ps.common.Program#1726043454198#0.6775744709437909
#> 17   fr.ird.referential.ps.common.Program#1742306567114#0.3279610697320161
#> 18   fr.ird.referential.ps.common.Program#1742306699358#0.7521250472635266
#> 19   fr.ird.referential.ps.common.Program#1742306748063#0.8144485417550063
#> 20  fr.ird.referential.ps.common.Program#1742307215135#0.10188296820927811
#> 21   fr.ird.referential.ps.common.Program#1742307270165#0.1423934943067212
#> 22   fr.ird.referential.ps.common.Program#1726494699541#0.8812018433671981
#> 23  fr.ird.referential.ps.common.Program#1726494644034#0.10598974977677977
#> 24   fr.ird.referential.ps.common.Program#1729643421733#0.5899283478668206
#> 25    fr.ird.referential.ps.common.Program#1726732972356#0.943823545843561
#> 26   fr.ird.referential.ps.common.Program#1742307315794#0.7310389385148582
#> 27   fr.ird.referential.ps.common.Program#1742307615221#0.2890639566367579
#> 28   fr.ird.referential.ps.common.Program#1742307682563#0.3072153441393439
#> 29  fr.ird.referential.ps.common.Program#1742307728624#0.22465347864570984
#>       code                                                  label1
#> 1   IRD-34    Stage 1 - Ongoing Abidjan data acquisitions (France)
#> 2   IRD-42    Stage 2 - Completed Dakar data acquisitions (France)
#> 3   IRD-40 Stage 2 - Completed Victoria data acquisitions (France)
#> 4   IRD-32                       AVDTH Atlantic 1999-2022 (France)
#> 5   IRD-37                           AVDTH Atlantic 2023+ (France)
#> 6   IRD-33                         AVDTH Indian 1999-2022 (France)
#> 7   IRD-38                             AVDTH Indian 2023+ (France)
#> 8   IRD-35      Stage 1 - Ongoing Dakar data acquisitions (France)
#> 9   SFA-80        Stage 1 - Ongoing Victoria acquisitions (Others)
#> 10  SFA-60         Stage 1 - Ongoing Victoria acquisitions (Spain)
#> 11  IRD-41  Stage 2 - Completed Abidjan data acquisitions (France)
#> 12  IRD-36   Stage 1 - Ongoing Victoria data acquisitions (France)
#> 13  SFA-50    Stage 1 - Ongoing Victoria acquisitions (Seychelles)
#> 14  SFA-61       Stage 2 - Completed Victoria acquisitions (Spain)
#> 15  SFA-91                                   AVDTH Indian (Others)
#> 16  SFA-51  Stage 2 - Completed Victoria acquisitions (Seychelles)
#> 17  IRD-92   Stage 3 - Ongoing Abidjan positions checking (France)
#> 18  IRD-93     Stage 3 - Ongoing Dakar positions checking (France)
#> 19  IRD-94  Stage 3 - Ongoing Victoria positions checking (France)
#> 20  IRD-95       Stage 4 - Ongoing Abidjan cross checking (France)
#> 21  IRD-96         Stage 4 - Ongoing Dakar cross checking (France)
#> 22  IEO-71                                    AVDTH Indian (Spain)
#> 23  SFA-70                               AVDTH Indian (Seychelles)
#> 24  IRD-90                         AVDTH Indian 1981-1998 (France)
#> 25  SFA-81      Stage 2 - Completed Victoria acquisitions (Others)
#> 26  IRD-97      Stage 4 - Ongoing Victoria cross checking (France)
#> 27  IRD-98        Stage 5 - Completed Abidjan validations (France)
#> 28  IRD-99          Stage 5 - Completed Dakar validations (France)
#> 29 IRD-100       Stage 5 - Completed Victoria validations (France)