R6 class samples creation

Create R6 reference object class samples

Super class

t3::list_t3 -> samples

Methods

Public methods

samples$new()

samples$add()

samples$filter_by_trip()

samples$clone()

Inherited methods

t3::list_t3$count()

t3::list_t3$extract()

t3::list_t3$filter()

t3::list_t3$name()

t3::list_t3$remove()

t3::list_t3$view()

Method new()

Initialize function for R6 samples class.

Usage

samples$new(...)

Arguments

...

(empty, list or R6-elementarysample classes) Nothing, a list of object R6-elementarysample classes or one object R6-elementarysample classes.

Method add()

Function for add a new elementarysample in the object samples.

Usage

samples$add(new_item)

Arguments

new_item

(list or R6-elementarysample classes) A list of object R6-elementarysample classes or one object R6-elementarysample classes.

Method filter_by_trip()

Function for filter elementarysample by trip identification.

Usage

samples$filter_by_trip(trip_id)

Arguments

trip_id

(character) Trip identification.

Method clone()

The objects of this class are cloneable with this method.

Usage

samples$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.