T3 process level 3

Level 3 of t3 process (modelislation).

t3_level3(
  inputs_level3,
  inputs_level3_path = NULL,
  periode_reference_level3 = NULL,
  target_year,
  period_duration,
  distance_maximum = as.integer(5),
  number_sets_maximum = as.integer(5),
  set_weight_minimum = as.integer(6),
  minimum_set_frequency = 0.1,
  vessel_id_ignored = NULL,
  num.trees = 1000L,
  mtry = 2L,
  min.node.size = 5,
  seed_number = 7L,
  small_fish_only = F,
  plot_sample = F,
  avdth_patch_coord = F,
  ci = FALSE,
  ci_type = "all",
  Nboot = 50,
  plot_predict = FALSE,
  log_file = FALSE,
  log_path = NULL,
  log_name = "t3_level3",
  output_path = NULL,
  output_format = "eu",
  new_directory = FALSE,
  integrated_process = FALSE
)

Arguments

inputs_level3

Object of type data.frame expected. Inputs of levels 3 (see function path to level 3).

inputs_level3_path

Object of type character expected. Path to the folder containing yearly data output of the level 1 and 2 (output of the function the path to level 3). If provide, replace the inputs_level3 object.

periode_reference_level3

Object of type integer expected. Year(s) period of reference for modelling estimation.

target_year

Object of type integer expected. Year of interest for the model estimation and prediction.Default value is current year -1.

period_duration

Object of type integer expected. number of years use for the modelling. The default value is 5

distance_maximum

Object of type integer expected. Maximum distance between all sets of a sampled well. By default 5.

number_sets_maximum

Object of type integer expected. Maximum number of sets allowed in mixture. By default 5.

set_weight_minimum

Object of type integer expected. Minimum set size considered. Remove smallest set for which sample could not be representative. By default 6 t.

minimum_set_frequency

Object of type numeric expected. Minimum threshold proportion of set in awell to be used for model training in the process. By default 0.1.

vessel_id_ignored

Object of type integer expected. Specify list of vessel(s) id(s) to be ignored in the model estimation and prediction .By default NULL.

num.trees

Object of type integer expected. Number of trees to grow. This should not be set to too small a number, to ensure that every input row gets predicted at least a few times. The default value is 1000.

mtry

Object of type integer expected. Number of variables randomly sampled as candidates at each split. The default value is 2.

min.node.size

Object of type numeric expected. Minimum size of terminal nodes. Setting this number larger causes smaller trees to be grown (and thus take less time).The default value is 5.

seed_number

Object of type integer expected. Set the initial seed for the modelling. The default value is 7.

small_fish_only

Object of type logical expected. Whether the model estimate proportion for small fish only (< 10 kg).

plot_sample

logical. Whether the sample figure is computed. Default value = F

avdth_patch_coord

parameter waiting for coordinate conversion patch from avdth database

ci

Object of type logical expected. Logical indicating whether confidence interval is computed. The default value is FALSE as it is a time consuming step.

ci_type

Type of confidence interval to compute. The default value is "all". Other options are "set" for ci on each set, "t1" for ci on nominal catch by species, "t1-fmod" for ci on nominal catch by species and fishing mode "t2" and "t2-fmod" for ci by 1 degree square and month. A vector of several ci option can be provided. ci_type are computed only if the ci parameter is TRUE.

Nboot

Object of type numeric expected. The number of bootstrap samples desired for the ci computation. The default value is 10.

plot_predict

Object of type logical expected. Logical indicating whether maps of catch at size have to be done.

log_file

Object of type logical expected. Initiation or not for log file creation. By default FALSE (no).

log_path

Object of type character expected. Path of the log file directory. By default NULL.

log_name

Object of type character expected. Name of the log file. By default "t3_level2".

output_path

Object of class character expected. Outputs path directory. By default NULL.

output_format

Object of class character expected. By default "eu". Select outputs format regarding European format (eu) or United States format (us).

new_directory

Object of class logical expected. Initiate a new outputs directory of use an existing one. By default NULL.

integrated_process

Object of class logical expected. Indicate if the process is integrated in another (like the one in the function "t3_process"). By default FALSE.