T3 process compilation

Run the t3 process, with the possibility to run all the process or stop.

t3_process(
  process = "all",
  data_source = "t3_db",
  db_con,
  log_file = FALSE,
  log_path = NULL,
  periode_reference,
  countries,
  oceans,
  sample_type,
  trips_selected = NULL,
  species_rf1 = as.integer(c(1, 2, 3, 4, 9, 11)),
  rf1_lowest_limit = 0.8,
  rf1_highest_limit = 1.2,
  sunrise_schema = "sunrise",
  sunset_schema = "sunset",
  maximum_lf_class = as.integer(500),
  threshold_rf_minus10 = as.integer(500),
  threshold_rf_plus10 = as.integer(500),
  threshold_frequency_rf_minus10 = as.integer(75),
  threshold_frequency_rf_plus10 = as.integer(75),
  threshold_rf_total = as.integer(250)
)

Arguments

process

Object of class character expected. Specify here if you want to run the whole process or just a part of it. By default "all". Check detail section below for more information.

data_source

Object of class character expected. Identification of data source. By default "t3_db" but you can switch to "avdth_db".

db_con

Database connection R object expected. Mandatory argument for data source "t3_db", "avdth_db" and "sql_query".

log_file

Object of class logical expected. Initiation or not for log file creation. By default FALSE (no).

log_path

Object of class character expected. Path of the log file directory. By default NULL.

periode_reference

Object of class integer expected. Year(s) of the reference period coded on 4 digits. By default NULL.

countries

Object of class character expected. ISO code on 3 letters related to one or more countries. By default NULL.

oceans

Object of class integer expected. Ocean(s) related to data coded on 1 digit. By default NULL.

sample_type

(integer) Sample type identification (landing, observer, ...). By default NULL.

trips_selected

Object of class character expected. Additional parameter only used with data source "t3_db". Use trip(s) identification(s) for selected trip(s) kept in the query (by periode of reference and countries). By default NULL.

species_rf1

Object of type integer expected. Specie(s) code(s) used for the RF1 process. By default 1 (YFT), 2 (SKJ), 3 (BET), 4 (ALB), 9 (MIX) and 11 (LOT).

rf1_lowest_limit

Object of type numeric expected. Verification value for the lowest limit of the RF1. By default 0.8.

rf1_highest_limit

Object of type numeric expected. Verification value for the highest limit of the RF1. By default 1.2.

sunrise_schema

Object of class character expected. Sunrise caracteristic. By default "sunrise" (top edge of the sun appears on the horizon). See function fishing_time() for more details.

sunset_schema

Object of class character expected. Sunset caracteristic. By default "sunset" (sun disappears below the horizon, evening civil twilight starts). See function fishing_time() for more details.

maximum_lf_class

Object of type integer expected. Theorical maximum lf class that can occur (all species considerated). By default 500.

threshold_rf_minus10

Object of type integer expected. Threshold limite value for raising factor on individuals category minus 10. By default 500.

threshold_rf_plus10

Object of type integer expected. Threshold limite value for raising factor on individuals category plus 10. By default 500.

threshold_frequency_rf_minus10

Object of type integer expected. Threshold limite frequency value for raising factor on individuals category minus 10. By default 75.

threshold_frequency_rf_plus10

Object of type integer expected. Threshold limite frequency value for raising factor on individuals category plus 10. By default 75.

threshold_rf_total

Object of type integer expected. Threshold limite value for raising factor (all categories). By default 250.

Details

For the argument "process", you can choose between 5 modalities (descending size classification):

all: argument by default, you launch all the process

level1: you launch data model initialisation and the process level 1

level2: you launch data model initialisation and the process level 2

until_level2: you launch data model initialisation, the process level 1 and 2