R6 class trip

Create R6 reference object class trip

Methods

Public methods

trip$new()

trip$clone()

Method new()

Initialize function for R6 trip class.

Usage

trip$new(
  trip_id,
  flag_code,
  departure_date,
  trip_end_date,
  logbook_availability_code,
  landing_well_content_code,
  vessel_code,
  vessel_type_code
)

Arguments

trip_id

Object of class character expected. Trip identification.

flag_code

Object of class character expected. Fleet identification.

departure_date

Object of class character expected. Departure date in format year month day hour minute second.

trip_end_date

Object of class character expected. Landing date in format ymd UTC.

logbook_availability_code

Object of class integer expected. Logbook availability value, 1 for available and 0 for not.

landing_well_content_code

Object of class integer expected. Informe if the fish hold empty at the end of the trip, 1 for yes and 0 for not.

vessel_code

Object of class integer expected. Vessel identification.

vessel_type_code

Object of class character expected. Vessel type identification.

Method clone()

The objects of this class are cloneable with this method.

Usage

trip$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.