R6 class full_trips creation

Create R6 reference object class full_trips

Super class

t3::list_t3 -> full_trips

Methods

Public methods

full_trips$create_full_trips()

full_trips$filter_by_periode()

full_trips$add_activities()

full_trips$add_elementarycatches()

full_trips$add_elementarylandings()

full_trips$add_wells_samples()

full_trips$rf1()

full_trips$rf2()

full_trips$conversion_weigth_category()

full_trips$set_count()

full_trips$set_duration()

full_trips$time_at_sea()

full_trips$fishing_time()

full_trips$searching_time()

full_trips$sample_number_measured_extrapolation()

full_trips$sample_length_class_ld1_to_lf()

full_trips$sample_length_class_step_standardisation()

full_trips$well_set_weigth_categories()

full_trips$standardised_sample_creation()

full_trips$standardised_sample_set_creation()

full_trips$raised_factors_determination()

full_trips$raised_standardised_sample_set()

full_trips$path_to_level3()

full_trips$data_preparatory()

full_trips$random_forest_models()

full_trips$models_checking()

full_trips$data_formatting_for_predictions()

full_trips$model_predictions()

full_trips$show_me_what_you_got()

full_trips$clone()

Inherited methods

t3::list_t3$add()

t3::list_t3$count()

t3::list_t3$extract()

t3::list_t3$filter()

t3::list_t3$initialize()

t3::list_t3$name()

t3::list_t3$remove()

t3::list_t3$view()

Method create_full_trips()

Creation of full trip item from trips.

Usage

full_trips$create_full_trips(object_trips)

Arguments

object_trips

(R6-trips) A R6 reference object of class trips.

Method filter_by_periode()

Function for filter full trips by a reference periode.

Usage

full_trips$filter_by_periode(periode_reference)

Arguments

periode_reference

(integer) Year(s) in 4 digits format.

Method add_activities()

Function for add activities in full trips object.

Usage

full_trips$add_activities(object_activities)

Arguments

object_activities

(R6-activities) A R6 reference object of class activities.

Method add_elementarycatches()

Function for add elementary catches in full trips object.

Usage

full_trips$add_elementarycatches(object_elementarycatches)

Arguments

object_elementarycatches

(R6-elementarycatches) A R6 reference object of class elementarycatches.

Method add_elementarylandings()

Function for add elementary landings in full trips object.

Usage

full_trips$add_elementarylandings(object_elementarylandings)

Arguments

object_elementarylandings

(R6-elementarylandings) A R6 reference object of class elementarylandings.

Method add_wells_samples()

Function for add wells and samples caracteristics in full trips object.

Usage

full_trips$add_wells_samples(object_wells)

Arguments

object_wells

(R6-wells) A R6 reference object of class wells.

Method rf1()

Process of Raising Factor level 1 (RF1).

Usage

full_trips$rf1(species_rf1, rf1_lowest_limit = 0.8, rf1_highest_limit = 1.2)

Arguments

species_rf1

Object of type integer expected. Specie(s) code(s) used for the RF1 process.

rf1_lowest_limit

Object of type numeric expected. Verification value for the lowest limit of the RF1. By default 0.8.

rf1_highest_limit

Object of type numeric expected. Verification value for the highest limit of the RF1. By default 1.2.

Method rf2()

Process of Raising Factor level 2 (rf2).

Usage

full_trips$rf2()

Method conversion_weigth_category()

Process of logbook weigth categories conversion.

Usage

full_trips$conversion_weigth_category()

Method set_count()

Process for postive sets count.

Usage

full_trips$set_count()

Method set_duration()

Process for set duration calculation (in hours).

Usage

full_trips$set_duration(set_duration_ref)

Arguments

set_duration_ref

(data frame) Data and parameters for set duration calculation (by year, country, ocean and school type).

Method time_at_sea()

Process for time at sea calculation (in hours).

Usage

full_trips$time_at_sea()

Method fishing_time()

Process for fishing time calculation (in hours).

Usage

full_trips$fishing_time(sunrise_schema = "sunrise", sunset_schema = "sunset")

Arguments

sunrise_schema

(character) Sunrise caracteristic. By default "sunrise" (top edge of the sun appears on the horizon). See below for more details.

sunset_schema

(character) Sunset caracteristic. By default "sunset" (sun disappears below the horizon, evening civil twilight starts). See below for more details.

Details

Available variables are:

"sunrise"sunrise (top edge of the sun appears on the horizon)

"sunriseEnd"sunrise ends (bottom edge of the sun touches the horizon)

"goldenHourEnd"morning golden hour (soft light, best time for photography) ends

"solarNoon"solar noon (sun is in the highest position)

"goldenHour"evening golden hour starts

"sunsetStart"sunset starts (bottom edge of the sun touches the horizon)

"sunset"sunset (sun disappears below the horizon, evening civil twilight starts)

"dusk"dusk (evening nautical twilight starts)

"nauticalDusk"nautical dusk (evening astronomical twilight starts)

"night"night starts (dark enough for astronomical observations)

"nadir"nadir (darkest moment of the night, sun is in the lowest position)

"nightEnd"night ends (morning astronomical twilight starts)

"nauticalDawn"nautical dawn (morning nautical twilight starts)

"dawn"dawn (morning nautical twilight ends, morning civil twilight starts)

Method searching_time()

Process for searching time calculation (in hours, fishing time minus sets durations).

Usage

full_trips$searching_time()

Method sample_number_measured_extrapolation()

Process for sample number measured individuals extrapolation to sample number individuals counted.

Usage

full_trips$sample_number_measured_extrapolation()

Method sample_length_class_ld1_to_lf()

Process for length conversion, if necessary, in length fork (lf). Furthermore, variable "sample_number_measured_extrapolated" of process 2.1 will converse in variable "sample_number_measured_extrapolated_lf" (Notably due to the creation of new lf classes during some conversions).

Usage

full_trips$sample_length_class_ld1_to_lf(length_step)

Arguments

length_step

(data.frame) Data frame object with length ratio between ld1 and lf class.

Method sample_length_class_step_standardisation()

Process for step standardisation of lf length class.

Usage

full_trips$sample_length_class_step_standardisation(
  maximum_lf_class = as.integer(500)
)

Arguments

maximum_lf_class

(integer) Theorical maximum lf class that can occur (all species considerated). By default 500.

Method well_set_weigth_categories()

Process for well set weigth categories definition.

Usage

full_trips$well_set_weigth_categories(sample_set)

Arguments

sample_set

(data.frame) Data frame object with weighted weigh of each set sampled.

Method standardised_sample_creation()

Object standardised sample creation.

Usage

full_trips$standardised_sample_creation()

Method standardised_sample_set_creation()

R6 object standardised sample set creation.

Usage

full_trips$standardised_sample_set_creation(length_weight_relationship_data)

Arguments

length_weight_relationship_data

(data.frame) Data frame object with parameters for length weight relationship.

Method raised_factors_determination()

Raised factors determination for weigth sample set to set.

Usage

full_trips$raised_factors_determination(
  threshold_rf_minus10 = as.integer(500),
  threshold_rf_plus10 = as.integer(500),
  threshold_frequency_rf_minus10 = as.integer(75),
  threshold_frequency_rf_plus10 = as.integer(75),
  threshold_rf_total = as.integer(250)
)

Arguments

threshold_rf_minus10

(integer) Threshold limite value for raising factor on individuals category minus 10. By default 500.

threshold_rf_plus10

(integer) Threshold limite value for raising factor on individuals category plus 10. By default 500.

threshold_frequency_rf_minus10

(integer) Threshold limite frequency value for raising factor on individuals category minus 10. By default 75.

threshold_frequency_rf_plus10

(integer) Threshold limite frequency value for raising factor on individuals category plus 10. By default 75.

threshold_rf_total

(integer) Threshold limite value for raising factor (all categories). By default 250.

Method raised_standardised_sample_set()

Application of process 2.8 raised factors on standardised sample set.

Usage

full_trips$raised_standardised_sample_set()

Method path_to_level3()

Temporary link to the R object model with Antoine D. modelisation process.

Usage

full_trips$path_to_level3()

Method data_preparatory()

Data preparatory for the t3 modelling process (level 3).

Usage

full_trips$data_preparatory(
  inputs_level3 = NULL,
  inputs_level3_path,
  outputs_directory,
  periode_reference = NULL,
  target_year = as.integer(lubridate::year(Sys.time() - 1)),
  period_duration = 4L,
  distance_maximum = as.integer(5),
  number_sets_maximum = as.integer(5),
  set_weight_minimum = as.integer(6),
  minimum_set_frequency = 0.1,
  vessel_id_ignored = NULL
)

Arguments

inputs_level3

(data frame) Imputs of levels 3 (see function path to level 3).

inputs_level3_path

(character) Path to the folder containing yearly data ouptut of the level 1 and 2.

outputs_directory

(character) Path of the t3 processes outputs directory.

periode_reference

(integer) Year(s) period of reference for modelling estimation.

target_year

(integer) Year of interest for the model estimaton and prediction.Default value is current year -1.

period_duration

(integer) number of year use for the modelling. The default value is 5

distance_maximum

(integer) Maximum distance between all sets of a sampled well. By default 5.

number_sets_maximum

(integer) Maximum number of sets allowed in mixture. By default 5.

set_weight_minimum

(integer) Minimum set size considered. Remove smallest set for which sample could not be representative. By default 6.

minimum_set_frequency

(numeric) Minimum freqency that a set could represent in a well. Another filter considering other set size in the well. By default 0.1.

vessel_id_ignored

(integer) Specify here vessel(s) id(s) if you whant to ignore it in the model estimation and prediction .By default NULL.

Method random_forest_models()

Modelling proportions in sets througth random forest models.

Usage

full_trips$random_forest_models(
  outputs_level3_process1,
  num.trees = 1000L,
  mtry = 2L,
  min.node.size = 5,
  seed_number = 7L
)

Arguments

outputs_level3_process1

(data frame) Output table data_lb_sample_screened from process 3.1.

num.trees

(integer) Number of trees to grow. This should not be set to too small a number, to ensure that every input row gets predicted at least a few times. The default value is 1000.

mtry

Number of variables randomly sampled as candidates at each split. The default value is 2.

min.node.size

Minimum size of terminal nodes. Setting this number larger causes smaller trees to be grown (and thus take less time).The default value is 5.

seed_number

Set the initial seed for the modelling. The default value is 7.

Method models_checking()

Load each full model and compute figures and tables to check the model quality. Furthermore, create a map of samples used for each model and relationship between logbook reports and samples.

Usage

full_trips$models_checking(outputs_level3_process2, outputs_path)

Arguments

outputs_level3_process2

(list) Outputs models and data from process 3.2.

outputs_path

(character) Outputs directory path.

Method data_formatting_for_predictions()

Formatting data for model predictions.

Usage

full_trips$data_formatting_for_predictions(
  inputs_level3,
  outputs_level3_process1,
  target_year,
  vessel_id_ignored = NULL
)

Arguments

inputs_level3

(data frame) Imputs of levels 3 (see function path to level 3).

outputs_level3_process1

(data frame) Output table data_lb_sample_screened from process 3.1.

target_year

(integer) The year of interest for the model estimaton and prediction.

vessel_id_ignored

(integer) Specify here vessel(s) id(s) if you whant to ignore it in the model estimation and prediction .By default NULL.

Method model_predictions()

Model predictions for the species composition and computing of catches.

Usage

full_trips$model_predictions(
  outputs_level3_process2,
  outputs_level3_process4,
  outputs_path,
  ci = FALSE,
  Nboot = 10,
  plot_predict = FALSE
)

Arguments

outputs_level3_process2

(list) Outputs from level 3 process 2 (random forest models).

outputs_level3_process4

(list) Outputs from level 3 process 4 (data formatting for predictions).

outputs_path

(character) Outputs directory path.

ci

Logical indicating whether confidence interval is computed. The default value is FALSE as it is a time consuming step.

Nboot

The number of bootstrap samples desired for the ci computation. The default value is 10.

plot_predict

Logical indicating whether maps of catch at size have to be done.

Method show_me_what_you_got()

Most powerfull and "schwifty" function in the univers for "open the T3 process" and manipulate in live R6 objects.

Usage

full_trips$show_me_what_you_got()

Method clone()

The objects of this class are cloneable with this method.

Usage

full_trips$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.