R6 class list_t3 creation

Create R6 reference object class list_t3.

Methods

Public methods

list_t3$new()

list_t3$name()

list_t3$add()

list_t3$remove()

list_t3$view()

list_t3$count()

list_t3$extract()

list_t3$filter()

list_t3$clone()

Method new()

Initialize function for R6 list_t3 class.

Usage

list_t3$new()

Method name()

Function for display item's name(s) attribut.

Usage

list_t3$name(attribut = "all")

Arguments

attribut

(character) Display all names with "all" attribut or specify one.

Method add()

Function for add new element in specific attribut.

Usage

list_t3$add(new_item, attribut = "data")

Arguments

new_item

(all type) Item to add.

attribut

(character) Attribut's name. By default "data".

Method remove()

Function for removed element of specific attribut.

Usage

list_t3$remove(item_id, attribut = "data")

Arguments

item_id

(numeric) Identification number of item to remove.

attribut

(character) Attribut's name. By default "data".

Method view()

Function for display element(s) of specific attribut.

Usage

list_t3$view(..., attribut = "data")

Arguments

...

(list) Identification number or name of item(s) to display.

attribut

(character) Attribut's name. By default "data".

Method count()

Function for display number of elements of a specific attribut.

Usage

list_t3$count(attribut = "data")

Arguments

attribut

(character) Attribut's name. By default "data".

Method extract()

Function for extract element(s) of a specific attribut.

Usage

list_t3$extract(attribut_l1 = "data", attribut_l2 = NULL, id = NULL)

Arguments

attribut_l1

(character) First strate attribut's name. By default "data".

attribut_l2

(character or NULL) Second strate attribut's name. By default NULL.

id

(integer) Number identification of element in the attribut.

Method filter()

Function for select attribut's element(s) by specific filter.

Usage

list_t3$filter(attribut_l1 = "data", filter)

Arguments

attribut_l1

(character) First strate attribut's name. By default "data".

filter

(character) Filter by a specific filter.

Method clone()

The objects of this class are cloneable with this method.

Usage

list_t3$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.