R6 class list_t3

Create R6 reference object class list_t3.

Methods

Public methods

list_t3$new()

list_t3$name()

list_t3$add()

list_t3$remove()

list_t3$view()

list_t3$count()

list_t3$extract()

list_t3$filter()

list_t3$filter_l1()

list_t3$extract_l1_element_value()

list_t3$modification_l1()

list_t3$clone()

Method new()

Initialize function for R6 list_t3 class.

Usage

list_t3$new()

Method name()

Function for display item's name(s) attribut.

Usage

list_t3$name(attribut = "all")

Arguments

attribut

(character) Display all names with "all" attribut or specify one.

Method add()

Function for add new element in specific attribut.

Usage

list_t3$add(new_item, attribut = "data", silent = TRUE)

Arguments

new_item

(all type) Item to add.

attribut

(character) Attribut's name. By default "data".

silent

logical expected. By default TRUE Display or not information when you run the process.

Method remove()

Function for removed element of specific attribut.

Usage

list_t3$remove(item_id, attribut = "data")

Arguments

item_id

(numeric) Identification number of item to remove.

attribut

(character) Attribut's name. By default "data".

Method view()

Function for display element(s) of specific attribut.

Usage

list_t3$view(..., attribut = "data")

Arguments

...

(list) Identification number or name of item(s) to display.

attribut

(character) Attribut's name. By default "data".

Method count()

Function for display number of elements of a specific attribut.

Usage

list_t3$count(attribut = "data")

Arguments

attribut

(character) Attribut's name. By default "data".

Method extract()

Function for extract element(s) of a specific attribut.

Usage

list_t3$extract(attribut_l1 = "data", attribut_l2 = NULL, id = NULL)

Arguments

attribut_l1

(character) First strate attribut's name. By default "data".

attribut_l2

(character or NULL) Second strate attribut's name. By default NULL.

id

(integer) Number identification of element in the attribut.

Method filter()

Function for select attribut's element(s) by specific filter.

Usage

list_t3$filter(attribut_l1 = "data", filter)

Arguments

attribut_l1

(character) First strate attribut's name. By default "data".

filter

(character) Filter by a specific filter.

Method filter_l1()

Function for select item(s) by specific selection.

Usage

list_t3$filter_l1(attribut_l1 = "data", filter, clone = FALSE)

Arguments

attribut_l1

Object of type character expected. First strate attribut's name. By default "data".

filter

Object of type character expected. Filter by specific selection. Use the pattern $path$ for specify path of element in the R6 object. For example: "$path$elementarycatch_id == "elementarycatch4" & $path$activity_id == "activity168""

clone

Object of type logical expected. TRUE if you want to create a new object (not link to the original object).

Method extract_l1_element_value()

Function for extract element(s) of a specific attribut.

Usage

list_t3$extract_l1_element_value(attribut_l1 = "data", element)

Arguments

attribut_l1

Object of type character expected. First strate attribut's name. By default "data".

element

Object of type character expected. Name of the element.

Method modification_l1()

Function for modification item(s).

Usage

list_t3$modification_l1(attribut_l1 = "data", modification, silent = TRUE)

Arguments

attribut_l1

Object of type character expected. First strate attribut's name. By default "data".

modification

Object of type character expected. Attribute to modify. Use the pattern $path$ for specify path of attribute in the R6 object. For example: "$path$activity_code = 1"

silent

Object of type logical expected. Display outputs of modificated values.

Method clone()

The objects of this class are cloneable with this method.

Usage

list_t3$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.