R6 class object_model_data creation

Create R6 reference object class object_model_data

Methods

Public methods

object_model_data$trips_object_creation()

object_model_data$activities_object_creation()

object_model_data$elementarycatches_object_creation()

object_model_data$elementarylandings_object_creation()

object_model_data$wells_object_creation()

object_model_data$setdurationrefs_data()

object_model_data$lengthsteps_data()

object_model_data$samplesets_data()

object_model_data$lengthweightrelationships_data()

object_model_data$clone()

Method trips_object_creation()

Creation of a R6 reference object class trips which contain one or more R6 reference object class trip.

Usage

object_model_data$trips_object_creation(
  data_source = "t3_db",
  db_con = NULL,
  periode_reference = NULL,
  countries = NULL,
  oceans = NULL,
  data_path = NULL,
  trips_selected = NULL,
  envir = NULL
)

Arguments

data_source

Object of class character expected. Identification of data source. By default "t3_db" but you can switch to "sql_query", "csv" (with separator character ";" and decimal ","), "rdata" or "envir" (for an object in the R environment).

db_con

Object of class "database" expected. Check dbConnect. An R's object which contain connection identifiers for a database. Necessary argument for data source "t3_db" and "sql_query".

periode_reference

Object of class integer expected. Year(s) of the reference period coded on 4 digits. Necessary argument for data source "t3_db". By default NULL.

countries

Object of class character expected. ISO code on 3 letters related to one or more countries. Necessary argument for data source "t3_db". By default NULL.

oceans

Object of class integer expected. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "t3_db". By default NULL.

data_path

Object of class character expected. Path of the data sql/csv/RData file. By default NULL.

trips_selected

Object of class character expected. Additional parameter only used with data source "t3_db". Use trip(s) identification(s) for selected trip(s) kept in the query (by periode of reference and countries). By default NULL.

envir

Object of class character expected. Specify an environment to look in for data source "envir". By default the first environment where data are found will be used.

Method activities_object_creation()

Creation of a R6 reference object class activities which contain one or more R6 reference object class activity.

Usage

object_model_data$activities_object_creation(
  data_source = "t3_db",
  db_con = NULL,
  periode_reference = NULL,
  countries = NULL,
  oceans = NULL,
  data_path = NULL,
  trips_selected = NULL,
  envir = NULL
)

Arguments

data_source

Object of class character expected. Identification of data source. By default "t3_db" but you can switch to "sql_query", "csv" (with separator character ";" and decimal ","), "rdata" or "envir" (for an object in the R environment).

db_con

Object of class "database" expected. Check dbConnect. An R's object which contain connection identifiers for a database. Necessary argument for data source "t3_db" and "sql_query".

periode_reference

Object of class integer expected. Year(s) of the reference period coded on 4 digits. Necessary argument for data source "t3_db". By default NULL.

countries

Object of class character expected. ISO code on 3 letters related to one or more countries. Necessary argument for data source "t3_db". By default NULL.

oceans

Object of class integer expected. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "t3_db". By default NULL.

data_path

Object of class character expected. Path of the data sql/csv/RData file. By default NULL.

trips_selected

Object of class character expected. Additional parameter only used with data source "t3_db". Use trip(s) identification(s) for selected trip(s) kept in the query (by periode of reference and countries). By default NULL.

envir

Object of class character expected. Specify an environment to look in for data source "envir". By default the first environment where data are found will be used.

Method elementarycatches_object_creation()

Creation of a R6 reference object class elementarycatches which contain one or more R6 reference object class elementarycatch.

Usage

object_model_data$elementarycatches_object_creation(
  data_source = "t3_db",
  db_con = NULL,
  periode_reference = NULL,
  countries = NULL,
  oceans = NULL,
  data_path = NULL,
  trips_selected = NULL,
  envir = NULL
)

Arguments

data_source

Object of class character expected. Identification of data source. By default "t3_db" but you can switch to "sql_query", "csv" (with separator character ";" and decimal ","), "rdata" or "envir" (for an object in the R environment).

db_con

Object of class "database" expected. Check dbConnect. An R's object which contain connection identifiers for a database. Necessary argument for data source "t3_db" and "sql_query".

periode_reference

Object of class integer expected. Year(s) of the reference period coded on 4 digits. Necessary argument for data source "t3_db". By default NULL.

countries

Object of class character expected. ISO code on 3 letters related to one or more countries. Necessary argument for data source "t3_db". By default NULL.

oceans

Object of class integer expected. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "t3_db". By default NULL.

data_path

Object of class character expected. Path of the data sql/csv/RData file. By default NULL.

trips_selected

Object of class character expected. Additional parameter only used with data source "t3_db". Use trip(s) identification(s) for selected trip(s) kept in the query (by periode of reference and countries). By default NULL.

envir

Object of class character expected. Specify an environment to look in for data source "envir". By default the first environment where data are found will be used.

Method elementarylandings_object_creation()

Creation of a R6 reference object class elementarylandings which contain one or more R6 reference object class elementarylanding

Usage

object_model_data$elementarylandings_object_creation(
  data_source = "t3_db",
  db_con = NULL,
  periode_reference = NULL,
  countries = NULL,
  oceans = NULL,
  data_path = NULL,
  trips_selected = NULL,
  envir = NULL
)

Arguments

data_source

Object of class character expected. Identification of data source. By default "t3_db" but you can switch to "sql_query", "csv" (with separator character ";" and decimal ","), "rdata" or "envir" (for an object in the R environment).

db_con

Object of class "database" expected. Check dbConnect. An R's object which contain connection identifiers for a database. Necessary argument for data source "t3_db" and "sql_query".

periode_reference

Object of class integer expected. Year(s) of the reference period coded on 4 digits. Necessary argument for data source "t3_db". By default NULL.

countries

Object of class character expected. ISO code on 3 letters related to one or more countries. Necessary argument for data source "t3_db". By default NULL.

oceans

Object of class integer expected. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "t3_db". By default NULL.

data_path

Object of class character expected. Path of the data sql/csv/RData file. By default NULL.

trips_selected

Object of class character expected. Additional parameter only used with data source "t3_db". Use trip(s) identification(s) for selected trip(s) kept in the query (by periode of reference and countries). By default NULL.

envir

Object of class character expected. Specify an environment to look in for data source "envir". By default the first environment where data are found will be used.

Method wells_object_creation()

Creation of a R6 reference object class wells which contain one or more R6 reference object class well, wellset and samples.

Usage

object_model_data$wells_object_creation(
  data_source = "t3_db",
  db_con = NULL,
  periode_reference = NULL,
  countries = NULL,
  oceans = NULL,
  sample_type = NULL,
  trips_selected = NULL,
  data_path_samples = NULL,
  data_path_wellplans = NULL,
  envir = NULL
)

Arguments

data_source

Object of class character expected. Identification of data source. By default "t3_db" but you can switch to "sql_query", "csv" (with separator character ";" and decimal ","), "rdata" or "envir" (for an object in the R environment).

db_con

Object of class "database" expected. Check dbConnect. An R's object which contain connection identifiers for a database. Necessary argument for data source "t3_db" and "sql_query".

periode_reference

Object of class integer expected. Year(s) of the reference period coded on 4 digits. Necessary argument for data source "t3_db". By default NULL.

countries

Object of class character expected. ISO code on 3 letters related to one or more countries. Necessary argument for data source "t3_db". By default NULL.

oceans

Object of class integer expected. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "t3_db". By default NULL.

sample_type

(integer) Sample type identification (landing, observer, ...). By default NULL.

trips_selected

Object of class character expected. Additional parameter only used with data source "t3_db". Use trip(s) identification(s) for selected trip(s) kept in the query (by periode of reference and countries). By default NULL.

data_path_samples

Object of class character expected. Path of the data sql/csv file for samples. By default NULL.

data_path_wellplans

Object of class character expected. Path of the data sql/csv file for well plans. By default NULL.

envir

Object of class character expected. Specify an environment to look in for data source "envir". By default the first environment where data are found will be used.

Method setdurationrefs_data()

Creation of a data frame object with parameters of set duration algorithms.

Usage

object_model_data$setdurationrefs_data(
  db_con = NULL,
  data_source = "t3_db",
  periode_reference = NULL,
  countries = NULL,
  data_path = NULL,
  envir = NULL
)

Arguments

db_con

Object of class "database" expected. Check dbConnect. An R's object which contain connection identifiers for a database. Necessary argument for data source "t3_db" and "sql_query".

data_source

Object of class character expected. Identification of data source. By default "t3_db" but you can switch to "sql_query", "csv" (with separator character ";" and decimal ","), "rdata" or "envir" (for an object in the R environment).

periode_reference

Object of class integer expected. Year(s) of the reference period coded on 4 digits. Necessary argument for data source "t3_db". By default NULL.

countries

Object of class character expected. ISO code on 3 letters related to one or more countries. Necessary argument for data source "t3_db". By default NULL.

data_path

Object of class character expected. Path of the data sql/csv/RData file. By default NULL.

envir

Object of class character expected. Specify an environment to look in for data source "envir". By default the first environment where data are found will be used.

Method lengthsteps_data()

Creation of a data frame object with length ratio between ld1 and lf class.

Usage

object_model_data$lengthsteps_data(
  db_con = NULL,
  data_source = "t3_db",
  data_path = NULL,
  envir = NULL
)

Arguments

db_con

Object of class "database" expected. Check dbConnect. An R's object which contain connection identifiers for a database. Necessary argument for data source "t3_db" and "sql_query".

data_source

Object of class character expected. Identification of data source. By default "t3_db" but you can switch to "sql_query", "csv" (with separator character ";" and decimal ","), "rdata" or "envir" (for an object in the R environment).

data_path

Object of class character expected. Path of the data sql/csv/RData file. By default NULL.

envir

Object of class character expected. Specify an environment to look in for data source "envir". By default the first environment where data are found will be used.

Method samplesets_data()

Creation of a data frame object with weighted weigth of each set sampled.

Usage

object_model_data$samplesets_data(
  data_source = "t3_db",
  db_con = NULL,
  periode_reference = NULL,
  countries = NULL,
  oceans = NULL,
  data_path = NULL,
  trips_selected = NULL,
  envir = NULL
)

Arguments

data_source

Object of class character expected. Identification of data source. By default "t3_db" but you can switch to "sql_query", "csv" (with separator character ";" and decimal ","), "rdata" or "envir" (for an object in the R environment).

db_con

Object of class "database" expected. Check dbConnect. An R's object which contain connection identifiers for a database. Necessary argument for data source "t3_db" and "sql_query".

periode_reference

Object of class integer expected. Year(s) of the reference period coded on 4 digits. Necessary argument for data source "t3_db". By default NULL.

countries

Object of class character expected. ISO code on 3 letters related to one or more countries. Necessary argument for data source "t3_db". By default NULL.

oceans

Object of class integer expected. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "t3_db". By default NULL.

data_path

Object of class character expected. Path of the data sql/csv/RData file. By default NULL.

trips_selected

Object of class character expected. Additional parameter only used with data source "t3_db". Use trip(s) identification(s) for selected trip(s) kept in the query (by periode of reference and countries). By default NULL.

envir

Object of class character expected. Specify an environment to look in for data source "envir". By default the first environment where data are found will be used.

Method lengthweightrelationships_data()

Creation of a data frame object with parameters for length weight relationship.

Usage

object_model_data$lengthweightrelationships_data(
  data_source = "t3_db",
  db_con = NULL,
  data_path = NULL,
  envir = NULL
)

Arguments

data_source

Object of class character expected. Identification of data source. By default "t3_db" but you can switch to "sql_query", "csv" (with separator character ";" and decimal ","), "rdata" or "envir" (for an object in the R environment).

db_con

Object of class "database" expected. Check dbConnect. An R's object which contain connection identifiers for a database. Necessary argument for data source "t3_db" and "sql_query".

data_path

Object of class character expected. Path of the data sql/csv/RData file. By default NULL.

envir

Object of class character expected. Specify an environment to look in for data source "envir". By default the first environment where data are found will be used.

Method clone()

The objects of this class are cloneable with this method.

Usage

object_model_data$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.