R6 class object_model_data

Create R6 reference object class object_model_data

Methods

Public methods

object_model_data$trips_object_creation()

object_model_data$activities_object_creation()

object_model_data$elementarycatches_object_creation()

object_model_data$elementarylandings_object_creation()

object_model_data$wells_object_creation()

object_model_data$setdurationrefs_data()

object_model_data$lengthsteps_data()

object_model_data$samplesets_data()

object_model_data$lengthweightrelationships_data()

object_model_data$clone()

Method trips_object_creation()

Creation of an R6 reference object class trips which contains one or more R6 reference object class trip.

Usage

object_model_data$trips_object_creation(
  data_source = "observe_database",
  database_connection = NULL,
  years_period = NULL,
  flag_codes = NULL,
  ocean_codes = NULL,
  vessel_type_codes = NULL,
  trip_ids = NULL,
  data_path = NULL,
  envir = NULL
)

Arguments

data_source

Object of class character expected. By default "observe_database". Identification of data source. You can switch between "observe_database", "avdth_database", "csv_file" (with separator ";" and decimal ","), "rdata_file" or "envir" (for an object in the R environment).

database_connection

Database connection R object expected. By default NULL. Mandatory argument for data source "observe_database" and "avdth_database".

years_period

Object of class integer expected. By default NULL. Year(s) of the reference time period coded on 4 digits. Mandatory for data source "observe_database" and "avdth_database".

flag_codes

Object of class character expected. By default NULL. Country(ies) code related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".

ocean_codes

Object of class integer expected. By default NULL. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "observe_database" and "avdth_database".

vessel_type_codes

Object of class integer expected. By default NULL. Vessel type(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".

trip_ids

Object of class character expected. By default NULL. Additional parameter only used with data source "observe_database". Use trip(s) identification(s) for selected trip(s) kept in the query. This argument overrides all others arguments like "years_period", "country" or "ocean".

data_path

Object of class character expected. By default NULL. Path of the data csv/RData file.

envir

Object of class character expected. By default the first environment where data are found will be used. Specify an environment to look in for data source "envir".

Method activities_object_creation()

Creation of a R6 reference object class activities which contain one or more R6 reference object class activity.

Usage

object_model_data$activities_object_creation(
  data_source = "observe_database",
  database_connection = NULL,
  years_period = NULL,
  flag_codes = NULL,
  ocean_codes = NULL,
  vessel_type_codes = NULL,
  trip_ids = NULL,
  data_path = NULL,
  envir = NULL
)

Arguments

data_source

Object of class character expected. By default "observe_database". Identification of data source. You can switch between "observe_database", "avdth_database", "csv_file" (with separator ";" and decimal ","), "rdata_file" or "envir" (for an object in the R environment).

database_connection

Database connection R object expected. By default NULL. Mandatory argument for data source "observe_database" and "avdth_database".

years_period

Object of class integer expected. By default NULL. Year(s) of the reference time period coded on 4 digits. Mandatory for data source "observe_database" and "avdth_database".

flag_codes

Object of class character expected. By default NULL. Country(ies) code related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".

ocean_codes

Object of class integer expected. By default NULL. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "observe_database" and "avdth_database".

vessel_type_codes

Object of class integer expected. By default NULL. Vessel type(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".

trip_ids

Object of class character expected. By default NULL. Additional parameter only used with data source "observe_database". Use trip(s) identification(s) for selected trip(s) kept in the query. This argument overrides all others arguments like "years_period", "country" or "ocean".

data_path

Object of class character expected. By default NULL. Path of the data csv/RData file.

envir

Object of class character expected. By default the first environment where data are found will be used. Specify an environment to look in for data source "envir".

Method elementarycatches_object_creation()

Creation of a R6 reference object class elementarycatches which contain one or more R6 reference object class elementarycatch.

Usage

object_model_data$elementarycatches_object_creation(
  data_source = "observe_database",
  database_connection = NULL,
  years_period = NULL,
  flag_codes = NULL,
  ocean_codes = NULL,
  vessel_type_codes = NULL,
  species_fate_codes = NULL,
  trip_ids = NULL,
  data_path = NULL,
  envir = NULL
)

Arguments

data_source

Object of class character expected. By default "observe_database". Identification of data source. You can switch between "observe_database", "avdth_database", "csv_file" (with separator ";" and decimal ","), "rdata_file" or "envir" (for an object in the R environment).

database_connection

Database connection R object expected. By default NULL. Mandatory argument for data source "observe_database" and "avdth_database".

years_period

Object of class integer expected. By default NULL. Year(s) of the reference time period coded on 4 digits. Mandatory for data source "observe_database" and "avdth_database".

flag_codes

Object of class character expected. By default NULL. Country(ies) code related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".

ocean_codes

Object of class integer expected. By default NULL. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "observe_database" and "avdth_database".

vessel_type_codes

Object of class integer expected. By default NULL. Vessel type(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".

species_fate_codes

Object of class integer expected. By default NULL. Specie fate(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".

trip_ids

Object of class character expected. By default NULL. Additional parameter only used with data source "observe_database". Use trip(s) identification(s) for selected trip(s) kept in the query. This argument overrides all others arguments like "years_period", "country" or "ocean".

data_path

Object of class character expected. By default NULL. Path of the data csv/RData file.

envir

Object of class character expected. By default the first environment where data are found will be used. Specify an environment to look in for data source "envir".

Method elementarylandings_object_creation()

Creation of a R6 reference object class elementarylandings which contain one or more R6 reference object class elementarylanding

Usage

object_model_data$elementarylandings_object_creation(
  data_source = "observe_database",
  database_connection = NULL,
  years_period = NULL,
  flag_codes = NULL,
  ocean_codes = NULL,
  vessel_type_codes = NULL,
  trip_ids = NULL,
  data_path = NULL,
  envir = NULL
)

Arguments

data_source

Object of class character expected. By default "observe_database". Identification of data source. You can switch between "observe_database", "avdth_database", "csv_file" (with separator ";" and decimal ","), "rdata_file" or "envir" (for an object in the R environment).

database_connection

Database connection R object expected. By default NULL. Mandatory argument for data source "observe_database" and "avdth_database".

years_period

Object of class integer expected. By default NULL. Year(s) of the reference time period coded on 4 digits. Mandatory for data source "observe_database" and "avdth_database".

flag_codes

Object of class character expected. By default NULL. Country(ies) code related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".

ocean_codes

Object of class integer expected. By default NULL. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "observe_database" and "avdth_database".

vessel_type_codes

Object of class integer expected. By default NULL. Vessel type(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".

trip_ids

Object of class character expected. By default NULL. Additional parameter only used with data source "observe_database". Use trip(s) identification(s) for selected trip(s) kept in the query. This argument overrides all others arguments like "years_period", "country" or "ocean".

data_path

Object of class character expected. By default NULL. Path of the data csv/RData file.

envir

Object of class character expected. By default the first environment where data are found will be used. Specify an environment to look in for data source "envir".

Method wells_object_creation()

Creation of a R6 reference object class wells which contain one or more R6 reference object class well, wellset, samples and elementarywellplan.

Usage

object_model_data$wells_object_creation(
  data_source = "observe_database",
  database_connection = NULL,
  years_period = NULL,
  flag_codes = NULL,
  ocean_codes = NULL,
  vessel_type_codes = NULL,
  sample_type_codes = NULL,
  trip_ids = NULL,
  data_path_sample = NULL,
  data_path_wellplan = NULL,
  envir = NULL
)

Arguments

data_source

Object of class character expected. By default "observe_database". Identification of data source. You can switch between "observe_database", "avdth_database", "csv_file" (with separator ";" and decimal ","), "rdata_file" or "envir" (for an object in the R environment).

database_connection

Database connection R object expected. By default NULL. Mandatory argument for data source "observe_database" and "avdth_database".

years_period

Object of class integer expected. By default NULL. Year(s) of the reference time period coded on 4 digits. Mandatory for data source "observe_database" and "avdth_database".

flag_codes

Object of class character expected. By default NULL. Country(ies) code related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".

ocean_codes

Object of class integer expected. By default NULL. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "observe_database" and "avdth_database".

vessel_type_codes

Object of class integer expected. By default NULL. Vessel type(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".

sample_type_codes

Object of class integer expected. By default NULL. Sample type identification.

trip_ids

Object of class character expected. By default NULL. Additional parameter only used with data source "observe_database". Use trip(s) identification(s) for selected trip(s) kept in the query. This argument overrides all others arguments like "years_period", "country" or "ocean".

data_path_sample

Object of class character expected. By default NULL. Path of the data sql/csv file for samples.

data_path_wellplan

Object of class character expected. By default NULL. Path of the data sql/csv file for well plans.

envir

Object of class character expected. By default NULL. Specify an environment to look in for data source "envir".

Method setdurationrefs_data()

Creation of a data frame object with parameters of set duration algorithms.

Usage

object_model_data$setdurationrefs_data(
  data_source = "csv_file",
  data_path = NULL,
  envir = NULL
)

Arguments

data_source

Object of class character expected. Identification of data source. By default "csv_file" (with separator character ";" and decimal ","). Identification of data source. You can switch to "rdata_file" or "envir" (for an object in the R environment).

data_path

Object of class character expected. By default NULL. Mandatory argument for data source "csv_file", "rdata_file" or "envir". Path of the data file.

envir

Object of class character expected. By default NULL. Specify an environment to look in for data source "envir".

Method lengthsteps_data()

Creation of a data frame object with length ratio between ld1 and lf class.

Usage

object_model_data$lengthsteps_data(
  data_source = "csv_file",
  data_path = NULL,
  envir = NULL
)

Arguments

data_source

Object of class character expected. By default "csv_file" (with separator character ";" and decimal ","). Identification of data source. You can switch to "rdata_file" or "envir" (for an object in the R environment).

data_path

Object of class character expected. By default NULL. Mandatory argument for data source "csv_file", "rdata_file" or "envir". Path of the data file.

envir

Object of class character expected. By default NULL. Specify an environment to look in for data source "envir".

Method samplesets_data()

Creation of a data frame object with weighted weigth of each set sampled.

Usage

object_model_data$samplesets_data(
  data_source = "observe_database",
  database_connection = NULL,
  years_period = NULL,
  flag_codes = NULL,
  ocean_codes = NULL,
  vessel_type_codes = NULL,
  trip_ids = NULL,
  data_path = NULL,
  envir = NULL
)

Arguments

data_source

Object of class character expected. By default "observe_database". Identification of data source. You can switch between "observe_database", "avdth_database", "csv_file" (with separator ";" and decimal ","), "rdata_file" or "envir" (for an object in the R environment).

database_connection

Database connection R object expected. By default NULL. Mandatory argument for data source "observe_database" and "avdth_database".

years_period

Object of class integer expected. By default NULL. Year(s) of the reference time period coded on 4 digits. Mandatory for data source "observe_database" and "avdth_database".

flag_codes

Object of class character expected. By default NULL. Country(ies) code related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".

ocean_codes

Object of class integer expected. By default NULL. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "observe_database" and "avdth_database".

vessel_type_codes

Object of class integer expected. By default NULL. Vessel type(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".

trip_ids

Object of class character expected. By default NULL. Additional parameter only used with data source "observe_database". Use trip(s) identification(s) for selected trip(s) kept in the query. This argument overrides all others arguments like "years_period", "country" or "ocean".

data_path

Object of class character expected. By default NULL. Path of the data csv/RData file.

envir

Object of class character expected. By default the first environment where data are found will be used. Specify an environment to look in for data source "envir".

Method lengthweightrelationships_data()

Creation of a data frame object with parameters for length weight relationship.

Usage

object_model_data$lengthweightrelationships_data(
  data_source = "observe_database",
  database_connection = NULL,
  data_path = NULL,
  envir = NULL
)

Arguments

data_source

Object of class character expected. By default "observe_database". Identification of data source. You can switch to "csv_file" (with separator character ";" and decimal ","), "rdata_file" or "envir" (for an object in the R environment).

database_connection

Database connection R object expected. By default NULL. Mandatory argument for data source "observe_database".

data_path

Object of class character expected. By default NULL. Path of the data file.

envir

Object of class character expected. By default NULL. Specify an environment to look in for data source "envir".

Method clone()

The objects of this class are cloneable with this method.

Usage

object_model_data$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.