T3 process level 1

Level 1 of t3 process (logbooks standardisation).

t3_level1(
  object_model_data,
  object_full_trips,
  rf1_computation = TRUE,
  species_fao_codes_rf1 = c("YFT", "SKJ", "BET", "ALB", "TUN", "MIX", "LOT"),
  species_fate_codes_rf1 = as.integer(c(6, 11)),
  vessel_type_codes_rf1 = as.integer(c(4, 5, 6)),
  rf1_lowest_limit = 0.8,
  rf1_highest_limit = 1.2,
  fishing_effort_computation = TRUE,
  sunrise_schema = "sunrise",
  sunset_schema = "sunset",
  log_file = FALSE,
  log_path = NULL,
  log_name = "t3_level1",
  output_path = NULL,
  output_format = "eu",
  new_directory = FALSE,
  integrated_process = FALSE,
  referential_template = "observe"
)

Arguments

object_model_data

Object of classes object_model_data and R6 expected.

object_full_trips

Object of classes full_trips, list_t3 and R6 expected.

rf1_computation

Object of class logical expected. If FALSE rf1 is not calculated (rf1=1 for all trips). By default TRUE, the rf1 is calculated for each trip.

species_fao_codes_rf1

Object of type character expected.Specie(s) FAO code(s) used for the RF1 process. By default, use codes YFT (*Thunnus albacares*), SKJ (*Katsuwonus pelamis*), BET (*Thunnus obesus*), ALB (*Thunnus alalunga*), LOT (*Thunnus tonggol*) and TUN/MIX (mix of tunas species in Observe/AVDTH database) (French and Mayotte fleets).

species_fate_codes_rf1

Object of type integer expected. By default 6 ("Retained, presumably destined for the cannery"). Specie(s) fate code(s) used for the RF1 process.

vessel_type_codes_rf1

Object of type integer expected. By default 4, 5 and 6. Vessel type(s).

rf1_lowest_limit

Object of type numeric expected. Verification value for the lowest limit of the RF1. By default 0.8.

rf1_highest_limit

Object of type numeric expected. Verification value for the highest limit of the RF1. By default 1.2.

fishing_effort_computation

Object of class logical expected. Calculation or not of fishing effort indicators (time at sea, fishing time, set duration and searching time). By default TRUE (yes).

sunrise_schema

Object of class character expected. Sunrise characteristic. By default "sunrise" (top edge of the sun appears on the horizon). See function fishing_time() for more details.

sunset_schema

Object of class character expected. Sunset characteristic. By default "sunset" (sun disappears below the horizon, evening civil twilight starts). See function fishing_time() for more details.

log_file

Object of class logical expected. Initiation or not for log file creation. By default FALSE (no).

log_path

Object of class character expected. Path of the log file directory. By default NULL.

log_name

Object of class character expected. Name of the log file. By default "t3_level1".

output_path

Object of class character expected. Outputs path directory. By default NULL.

output_format

Object of class character expected. By default "eu". Select outputs format regarding European format (eu) or United States format (us).

new_directory

Object of class logical expected. Initiate a new outputs directory of use an existing one. By default NULL.

integrated_process

Object of class logical expected. Indicate if the process is integrated in another (like the one in the function "t3_process"). By default FALSE.

referential_template

Object of class character expected. By default "observe". Referential template selected (for example regarding the activity_code). You can switch to "avdth".

Value

The function a R6 reference object of class "object_full_trips".