T3 process level 2

Level 2 of t3 process (samples standardisation).

t3_level2(
  object_model_data,
  object_full_trips,
  maximum_lf_class = as.integer(500),
  threshold_rf_minus10 = as.integer(500),
  threshold_rf_plus10 = as.integer(500),
  threshold_frequency_rf_minus10 = as.integer(75),
  threshold_frequency_rf_plus10 = as.integer(75),
  threshold_rf_total = as.integer(250),
  log_file = FALSE,
  log_path = NULL,
  log_name = "t3_level2",
  output_path = NULL,
  output_format = "eu",
  new_directory = FALSE,
  integrated_process = FALSE,
  referential_template = "observe"
)

Arguments

object_model_data

Object of classes object_model_data and R6 expected.

object_full_trips

Object of classes full_trips, list_t3 and R6 expected.

maximum_lf_class

Object of type integer expected. Theorical maximum lf class that can occur (all species considerated). By default 500.

threshold_rf_minus10

Object of type integer expected. Threshold limite value for raising factor on individuals category minus 10. By default 500.

threshold_rf_plus10

Object of type integer expected. Threshold limite value for raising factor on individuals category plus 10. By default 500.

threshold_frequency_rf_minus10

Object of type integer expected. Threshold limite frequency value for raising factor on individuals category minus 10. By default 75.

threshold_frequency_rf_plus10

Object of type integer expected. Threshold limite frequency value for raising factor on individuals category plus 10. By default 75.

threshold_rf_total

Object of type integer expected. Threshold limite value for raising factor (all categories). By default 250.

log_file

Object of type logical expected. Initiation or not for log file creation. By default FALSE (no).

log_path

Object of type character expected. Path of the log file directory. By default NULL.

log_name

Object of type character expected. Name of the log file. By default "t3_level2".

output_path

Object of class character expected. Outputs path directory. By default NULL.

output_format

Object of class character expected. By default "eu". Select outputs format regarding European format (eu) or United States format (us).

new_directory

Object of class logical expected. Initiate a new outputs directory of use an existing one. By default NULL.

integrated_process

Object of class logical expected. Indicate if the process is integrated in another (like the one in the function "t3_process"). By default FALSE.

referential_template

Object of class character expected. By default "observe". Referential template selected (for example regarding the activity_code). You can switch to "avdth".

Value

The function a R6 reference object of class "object_full_trips".