T3 process compilation

Run the t3 process, with the possibility to run all the process or stop.

t3_process(
  process = "all",
  data_source = "observe_database",
  database_connection,
  log_file = FALSE,
  log_path = NULL,
  output_path = NULL,
  output_format = "eu",
  new_directory = TRUE,
  years_period,
  flag_codes,
  ocean_codes,
  vessel_type_codes,
  species_fate_codes,
  sample_type_codes,
  trip_ids = NULL,
  species_fao_codes_rf1 = c("YFT", "SKJ", "BET", "ALB", "MIX", "LOT"),
  species_fate_codes_rf1 = as.integer(c(6, 11)),
  vessel_type_codes_rf1 = as.integer(c(4, 5, 6)),
  rf1_lowest_limit = 0.8,
  rf1_highest_limit = 1.2,
  sunrise_schema = "sunrise",
  sunset_schema = "sunset",
  maximum_lf_class = as.integer(500),
  threshold_rf_minus10 = as.integer(500),
  threshold_rf_plus10 = as.integer(500),
  threshold_frequency_rf_minus10 = as.integer(75),
  threshold_frequency_rf_plus10 = as.integer(75),
  threshold_rf_total = as.integer(250),
  periode_reference_level3 = NULL,
  target_year,
  period_duration,
  distance_maximum = as.integer(5),
  number_sets_maximum = as.integer(5),
  set_weight_minimum = as.integer(6),
  minimum_set_frequency = 0.1,
  vessel_id_ignored = NULL,
  num.trees = 1000L,
  mtry = 2L,
  min.node.size = 5,
  seed_number = 7L,
  small_fish_only = FALSE,
  plot_sample = FALSE,
  avdth_patch_coord = FALSE,
  ci = FALSE,
  ci_type = "all",
  Nboot = 50,
  plot_predict = FALSE
)

Arguments

process

Object of class character expected. Specify here if you want to run the whole process or just a part of it. By default "all". Check detail section below for more information.

data_source

Object of class character expected. By default "observe_database". Identification of data source. You can to "avdth_database".

database_connection

Database connection, list of one or more R object(s) expected. By default NULL. Mandatory argument for data source "observe_database" ("PostgreSQLConnection" R object), corresponding to the second element of the object returned by `furdeb::postgresql_dbconnection()`. Or mandatory argument for data source"avdth_database" ("JDBCConnection" R object) corresponding to the second element of the object returned by `furdeb::access_dbconnection()`. For data source "observe_database", a list of "PostgreSQLConnection" R objects can be specified to query data from different observe databases. For example, a list of two database connection arguments for "observe_main" and "observe_acquisition" can be specified to simultaneously import and process recent data from acquisition database, which has not yet been imported into the main database, and older data from the main database.

log_file

Object of class logical expected. Initiation or not for log file creation. By default FALSE (no).

log_path

Object of class character expected. Path of the log file directory. By default NULL.

output_path

Object of class character expected. Outputs path directory. By default NULL.

output_format

Object of class character expected. By default "eu". Select outputs format regarding European format (eu) or United States format (us).

new_directory

Object of class logical expected. Initiate a new outputs directory of use an existing one. By default NULL.

years_period

Object of class integer expected. By default NULL. Year(s) of the reference time period coded on 4 digits. Mandatory for data source "observe_database" and "avdth_database".

flag_codes

Object of class character expected. By default NULL. Three letters country(ies) FAO code(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".

ocean_codes

Object of class integer expected. By default NULL. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "observe_database" and "avdth_database".

vessel_type_codes

Object of class integer expected. By default NULL. Vessel type(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".

species_fate_codes

Object of class integer expected. By default NULL. Specie fate(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".

sample_type_codes

Object of class integer expected. By default NULL. Sample type identification.

trip_ids

Object of class character expected. By default NULL. Additional parameter only used with data source "observe_database". Use trip(s) identification(s) for selected trip(s) kept in the query. This argument overrides all others arguments like "years_period", "country" or "ocean".

species_fao_codes_rf1

Object of type character expected. By default YFT, SKJ, BET, ALB, MIX and LOT. Specie(s) FAO code(s) used for the RF1 process.

species_fate_codes_rf1

Object of type integer expected. By default 6 ("Retained, presumably destined for the cannery"). Specie(s) fate code(s) used for the RF1 process.

vessel_type_codes_rf1

Object of type integer expected. By default 4, 5 and 6. Vessel type(s).

rf1_lowest_limit

Object of type numeric expected. Verification value for the lowest limit of the RF1. By default 0.8.

rf1_highest_limit

Object of type numeric expected. Verification value for the highest limit of the RF1. By default 1.2.

sunrise_schema

Object of class character expected. Sunrise characteristic. By default "sunrise" (top edge of the sun appears on the horizon). See function fishing_time() for more details.

sunset_schema

Object of class character expected. Sunset characteristic. By default "sunset" (sun disappears below the horizon, evening civil twilight starts). See function fishing_time() for more details.

maximum_lf_class

Object of type integer expected. Theorical maximum lf class that can occur (all species considerated). By default 500.

threshold_rf_minus10

Object of type integer expected. Threshold limit value for raising factor on individuals category minus 10. By default 500.

threshold_rf_plus10

Object of type integer expected. Threshold limit value for raising factor on individuals category plus 10. By default 500.

threshold_frequency_rf_minus10

Object of type integer expected. Threshold limit frequency value for raising factor on individuals category minus 10. By default 75.

threshold_frequency_rf_plus10

Object of type integer expected. Threshold limit frequency value for raising factor on individuals category plus 10. By default 75.

threshold_rf_total

Object of type integer expected. Threshold limit value for raising factor (all categories). By default 250.

periode_reference_level3

Object of type integer expected. Year(s) period of reference for modelling estimation.

target_year

Object of type integer expected. Year of interest for the model estimation and prediction.Default value is current year -1.

period_duration

Object of type integer expected. number of years use for the modelling. The default value is 5

distance_maximum

Object of type integer expected. Maximum distance between all sets of a sampled well. By default 5.

number_sets_maximum

Object of type integer expected. Maximum number of sets allowed in mixture. By default 5.

set_weight_minimum

Object of type integer expected. Minimum set size considered. Remove smallest set for which sample could not be representative. By default 6 t.

minimum_set_frequency

Object of type numeric expected. Minimum threshold proportion of set in a well to be used for model training in the process. By default 0.1.

vessel_id_ignored

Object of type integer expected. Specify list of vessel(s) id(s) to be ignored in the model estimation and prediction .By default NULL.

num.trees

Object of type integer expected. Number of trees to grow. This should not be set to too small a number, to ensure that every input row gets predicted at least a few times. The default value is 1000.

mtry

Object of type integer expected. Number of variables randomly sampled as candidates at each split. The default value is 2.

min.node.size

Object of type numeric expected. Minimum size of terminal nodes. Setting this number larger causes smaller trees to be grown (and thus take less time).The default value is 5.

seed_number

Object of type integer expected. Set the initial seed for the modelling. The default value is 7.

small_fish_only

Object of type logical expected. Whether the model estimate proportion for small fish only (< 10 kg).

plot_sample

logical. Whether the sample figure is computed. Default value = F

avdth_patch_coord

parameter waiting for coordinate conversion patch from avdth database

ci

Object of type logical expected. Logical indicating whether confidence interval is computed. The default value is FALSE as it is a time consuming step.

ci_type

Type of confidence interval to compute. The default value is "all". Other options are "set" for ci on each set, "t1" for ci on nominal catch by species, "t1-fmod" for ci on nominal catch by species and fishing mode "t2" and "t2-fmod" for ci by 1 degree square and month. A vector of several ci option can be provided. ci_type are computed only if the ci parameter is TRUE.

Nboot

Object of type numeric expected. The number of bootstrap samples desired for the ci computation. The default value is 10.

plot_predict

Object of type logical expected. Logical indicating whether maps of catch at size have to be done.

Details

For the argument "process", you can choose between 4 modalities (descending size classification):

all: argument by default, you launch all the process

level1: you launch data model initialisation and the process level 1

level2: you launch data model initialisation and the process level 2

until_level2: you launch data model initialisation, the process level 1 and 2